#108 library upgrades, java 1.8 as min, removed deprecated methods, test improvements, etc
